+++ NEWS.md
@@ -1,5 +1,7 @@
 # forcats (development version)
 
+* `fct_collapse()` can now use `other_level = NA` (#291).
+
 * `fct_lump_n()` and `fct_lump_prop()` will now create an "Other" level even
   if it only consists of a single level. This makes them consistent with the
   other `fct_lump_*` functions (#274).